+++ ep-support-portal/src/app/app-routing.module.ts
@@ -2,15 +2,14 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { UserDetailComponent } from './pages/users/user-detail/user-detail.component';
-import { ReportDetailComponent } from './pages/reports/report-detail/report-detail.component';
-import { CompanyDetailComponent } from './pages/companies/company-detail/company-detail.component';
+import { ReportDetailComponent } from './pages/reports/report-detail/report-detail.component'
 import { PageNotFoundComponent } from './not-found.component';
 
 const appRoutes: Routes = [
+  { path: 'user-detail', component: UserDetailComponent },
   { path: 'report-detail', component: ReportDetailComponent },
-  { path: 'company-detail', component: CompanyDetailComponent },
-  { path: '', redirectTo: '/company-detail', pathMatch: 'full'},
-  { path: '**', component: PageNotFoundComponent }
+  { path: '', redirectTo: '/dashboard', pathMatch: 'full'},
+  { path: '**', component: PageNotFoundComponent },
 ];
 
 @NgModule({
