+++ ep-support-portal/src/app/app.component.html
@@ -14,12 +14,6 @@
     {{title}}!!
   </h1>
   <div class="col-sm-10 col-sm-offset-1">
-    <router-outlet></router-outlet>
-  </div>
-  <br>
-  <hr>
-  <br>
-  <div class="col-sm-10 col-sm-offset-1">
-    <!--<app-company></app-company>-->
+    <app-company></app-company>
   </div>
 </div>
