+++ ep-support-portal/src/app/pages/users/user-detail/user-detail.component.html
@@ -1,6 +1,3 @@
-<div *ngIf="user">
-  <h2>{{user.firstName}} details!</h2>
-  <div>
-    <label>id: </label>{{user.id}}
-  </div>
-</div>
+<p>
+  user-detail works!
+</p>
