+++ ep-support-portal/src/app/pages/users/user-detail/user-detail.component.ts
@@ -1,12 +1,15 @@
-import { Component, Input } from '@angular/core';
-
-import { User } from '././user'
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-user-detail',
   templateUrl: './user-detail.component.html',
   styleUrls: ['./user-detail.component.css']
 })
-export class UserDetailComponent {
-  @Input() user: User;
+export class UserDetailComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
 }
