--- ep-support-portal/src/app/pages/users/users.module.ts
@@ -1,5 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import { UsersListComponent } from './users-list.component'
